BUG: Correct python 2.6 bug for tests.
Moved is_vpt and to_vpt test under pyart.util.tests
